@@ -6,14 +6,17 @@
     "tsc": "tsc"
   },
   "dependencies": {
-    "@types/express": "^4.17.16",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "express": "^4.18.2",
     "http-errors": "~1.6.3",
     "jade": "^0.29.0",
     "morgan": "~1.9.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^1.8.10"
+    "ts-node": "^10.9.1"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.16",
+    "@types/node": "^18.11.18",
+    "typescript": "^4.9.5"
   }
 }
