@@ -1,11 +1,12 @@
-import express = require('express')
-const app = express()
-const port = 3000
+import express, { Express, Request, Response } from 'express';
 
-app.get('/', (req, res) => {
-  res.send('Hello World!')
-})
+const app: Express = express();
+const port = '3000';
+
+app.get('/', (req: Request, res: Response) => {
+  res.send('Express + TypeScript Server');
+});
 
 app.listen(port, () => {
-  console.log(`Example app listening on port ${port}`)
-})
+  console.log(`⚡️[server]: Server is running at http://localhost:${port}`);
+});
